move JsonNode to utils
