+++ rlink/src/runtime/coordinator/server.rs
@@ -8,7 +8,8 @@ use rand::prelude::*;
 use crate::api::checkpoint::Checkpoint;
 use crate::api::cluster::MetadataStorageType;
 use crate::api::cluster::{ResponseCode, StdResponse};
-use crate::dag::{DagManager, JsonDag};
+use crate::dag::utils::JsonDag;
+use crate::dag::DagManager;
 use crate::runtime::coordinator::checkpoint_manager::CheckpointManager;
 use crate::runtime::TaskManagerStatus;
 use crate::storage::metadata::MetadataStorage;
